+++ src/components/App/index.tsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import HandIcon from '@components/HandIcon';
 import PlayArea from '@components/PlayArea';
 import Style from './style';
 
@@ -7,6 +8,7 @@ const App = (): JSX.Element => {
   return (
     <Style className='app' onClick={() => setActiveCard(-1)}>
       <PlayArea activeCard={activeCard} setActiveCard={setActiveCard} />
+      <HandIcon />
     </Style>
   );
 };