+++ src/components/App/style.tsx
@@ -1,7 +1,6 @@
 import styled from 'styled-components';
 
 const Style = styled.div`
-  border: 1px solid red;
   color: #333;
   font-family: Arial;
   font-size: 14px;